Progress on filtering for the recommender system
This commit involves work on taking the user last 100 tweets, then creates a term-frequency document from this list. A top-ten-terms list if then derived from these term frequencies. An executor (For different forms of execution) and tweetcase (to hold sets of tweets) was also added. Lastly, a test file was added to my sandbox.