Changed Recommender System to no longer interact with Twitters API
This change was made to both the backend and api directories to keep things consistent. This version of the Recommender System is called "RecommenderTextual", to allow us to switch back to the old "Recommender" system if we need to urgently. An urgent change may be necessary due to an upcoming demo.