Added "How Many Days Ago" parameter functionality for generate method
This functionality limits how many days back to generate tweets from. For example, given a weeks worth a tweets, if this parameter is given the number "3", it will only make recommendations based on tweets from the past three days in the dataset it is given. Currently this functionality only accepts parameters from 1 to 7, which requires revision (Possibly changing this to hours, rather than days).