Backend: Fixed syntax error back, removed Rollbar code
Removed the Rollbar code from the main method because I'm not sure Rollbar can actually catch syntax errors. They're really not supposed to make it in there anyway but it's a bit of a hassle to figure out what went wrong. We have manually build a container, run it and observe what is wrong.