Added Extensive Stop-word list, and author/tweet feedback incorporation.
Added a new stop-word class with a huge list of english word and functionality for the "More from this author" and liking/disliking tweets. Bear in mind that the liking/disliking of tweets is not 100% implemented yet. So far, it just assumes that the user likes "Ruby" and dislikes "Java", since we don't have the machanism for receiving this data yet.