Added images to readme, preparing to make repository public
In preparation to make the repository public these changes are made to show a pleasing front for people to glance over our work. It's not finished yet, a gif of our app remains.

Not really proud of this HTML in the readme. But since we will be more or less parking this repository for the foreseeable future I think it's quite all right to do a bit of hacks for slightly more pleasing/compact display.